Correction on format
Change the numbering format (from "-f" to [math]::round), so that the
return is still a double and not a string.
On systems considering “,” as a decimal separator, the previous result
(string with a “,”) would round incorrectly the result when used with
[int] expression.
Exemple: “8,5” would translate to [int]85, leading to incorrect results.